renamed previous BillSplit.jsx to AddFriends.jsx and added AddFriends route
